feat: Implemented the get_closest_waypoint method of the TLDetector. Splitted the detector into a dummy and live detector.
